add typescript config
