+++ main.ts
@@ -1,0 +1,19 @@
+import version from './src/version'
+import {appTs} from "./src";
+
+const moduleWorker = async () => {
+    try {
+        console.log(`Crypto Module ${version}\n`);
+
+    } catch (error) {
+        console.error(error);
+    }
+}
+
+moduleWorker();
+appTs()
+
+console.log(1111111)
+console.log(1111111555)
+console.log(666)
+console.log(666677776)
