+++ src/index.js
@@ -1,0 +1,8 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.appTs = void 0;
+const appTs = () => console.log('log in ts mode////');
+exports.appTs = appTs;
+console.log(222);
+(0, exports.appTs)();
+//# sourceMappingURL=index.js.map
