+++ src/version.js
@@ -1,1 +1,4 @@
-export default '1.0.0';
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.default = '1.0.0';
+//# sourceMappingURL=version.js.map
