+++ src/version.ts
@@ -1,0 +1,1 @@
+export default '1.0.0';
